Major update

+++ commands/client/stop.js
@@ -24,11 +24,24 @@ module.exports = {
         const reason = interaction.options.getString('reason');
 
         //Checks
+        if(!interaction.member.roles.cache.find(role => role.name = REQUIRED_ROLE)) {
+            const error_permissions = new MessageEmbed()
+                .setColor('#ff2020')
+                .setThumbnail(`${interaction.member.user.displayAvatarURL({dynamic: true, size: 32})}`)
+                .setTitle('PermissionError')
+                .setDescription("I'm sorry but you do not have the permissions to perform this command. Please contact the server administrators if you believe that this is an error.")
+                .setFooter({text: `You need at least the '${REQUIRED_ROLE}' role to use this command.`});
+
+            interaction.reply({embeds: [error_permissions], ephemeral: is_ephemeral});
+            return;
+        }
 
         //Code
-        console.log("stop");
-        client.destroy();
-        //interaction.reply({content: "This command is currently unavailable.", ephemeral: is_ephemeral});
+        interaction.reply({content: "This command is under developpement. The bot will stop after this message.", ephemeral: false})
+            .then((messageResult) => {
+                console.log("stop");
+                client.destroy();
+            })
 
         /*Add buttons and ask for a confirmation.*/
     }
